+++ gpxviewr/baseweb/models/gpx_file.py
@@ -2,9 +2,7 @@ from pathlib import Path
 import secrets
 import time
 import os
-import geojson
 import json
-import phonenumbers
 from geopy import distance as geopy_distance
 from datetime import timedelta
 
@@ -14,23 +12,19 @@ from django.core.validators import URLValidator
 from django.core.exceptions import ValidationError
 from django.conf import settings
 from django.contrib.gis.db import models
-from django.core.exceptions import ObjectDoesNotExist
 from django.contrib.gis.measure import Distance
 from django.contrib.gis.db.models.functions import Distance as FDistance
 from django.contrib.gis.geos import Point
 from django.urls import reverse
 from django.db.models.signals import post_delete, pre_delete
-from django.db.models import Sum
 from django.dispatch import receiver
 
 from django_extensions.db.models import TimeStampedModel
 
 import overpy
 
-import numpy
 import pandas
-from gpx import GPX, Waypoint, Track
-from gpx.track_segment import TrackSegment
+from gpx import GPX, Waypoint
 from gpx.waypoint import Link
 
 
@@ -97,9 +91,13 @@ class GPXFile(TimeStampedModel):
         return False
 
     def get_all_segments_primary_keys(self):
+        from .gpx_track_segment import GPXTrackSegment
+
         return GPXTrackSegment.objects.all().filter(track__in=self.tracks.all().values_list('pk'))
 
     def get_waypoint_types(self) -> list:
+        from .gpx_way_point_type import GPXWayPointType
+
         data = []
         for wtp in GPXWayPointType.objects.all():
             o = self.wpt_options.get(wtp.name, {})
@@ -110,6 +108,8 @@ class GPXFile(TimeStampedModel):
         return data
 
     def get_waypoint_types_with_entries(self):
+        from .gpx_way_point_type import GPXWayPointType
+
         wtps = []
         for wtp in GPXWayPointType.objects.all():
             wtps.append({
@@ -208,7 +208,8 @@ class GPXFile(TimeStampedModel):
                     time.sleep(1)
 
     def query_data_osm(self, points, around_meters, point_type, around_duplicate) -> None:
-        from .tasks import gpx_waypoint_find_route_from_track
+        from baseweb.tasks import gpx_waypoint_find_route_from_track
+        from baseweb.models import GPXTrackWayPoint, GPXTrackSegmentPoint
 
         track_points = pandas.DataFrame(points, columns=["lat", "lon"])
         track_latlong_flatten = ",".join(track_points.to_numpy().flatten().astype("str"))
@@ -286,6 +287,7 @@ class GPXFile(TimeStampedModel):
         return None
 
     def load_file_to_database(self) -> None:
+        from baseweb.models import GPXTrackSegmentPoint, GPXTrack, GPXTrackSegment
         json_data = []
         gpxfile = GPX.from_file(self.file.path)
 
@@ -426,430 +428,6 @@ class GPXFile(TimeStampedModel):
             json_data.append(track_data)
 
         Path(self.get_gpx_json_filename()).write_text(json.dumps(json_data))
-
-
-class GPXTrack(TimeStampedModel):
-    gpx_file = models.ForeignKey("GPXFile", on_delete=models.CASCADE, related_name='tracks')
-    name = models.CharField(max_length=200, null=False, blank=False)
-    distance = models.FloatField()
-    link = models.URLField(max_length=6000, null=True)
-
-    def __str__(self) -> str:
-        return f"GPX Track {self.name}"
-
-    class Meta:
-        ordering = ['name',]
-
-    def get_human_distance(self) -> str:
-        d = int(self.distance / 1000)
-        return f"{d} km"
-
-    def get_total_ascent(self) -> int:
-        d = self.segments.all().aggregate(ascent=Sum('total_ascent'))
-
-        if d.get('ascent', None):
-            return int(d.get('ascent', None))
-
-        return None
-
-    def get_total_descent(self) -> int:
-        d = self.segments.all().aggregate(descent=Sum('total_descent'))
-
-        if d.get('descent', None):
-            return int(d.get('descent', None))
-
-        return None
-
-
-class GPXTrackSegment(TimeStampedModel):
-    number = models.IntegerField(null=False, blank=False, default=0)
-    track = models.ForeignKey("GPXTrack", on_delete=models.CASCADE, related_name='segments')
-    distance = models.FloatField()
-    total_ascent = models.FloatField(null=True, blank=False)
-    total_descent = models.FloatField(null=True, blank=False)
-
-    def __str__(self) -> str:
-        return f"GPX Track {self.track.name} Segment {self.number}"
-
-    class Meta:
-        ordering = ['track__name', 'number',]
-
-    def get_human_distance(self) -> str:
-        d = int(self.distance / 1000)
-        return f"{d} km"
-
-
-class GPXTrackSegmentPoint(TimeStampedModel):
-    number = models.BigIntegerField(null=True, db_index=True)
-    segment = models.ForeignKey("GPXTrackSegment", on_delete=models.CASCADE, related_name='points')
-    location = models.PointField(db_index=True)
-    elevation = models.FloatField(null=True, blank=False)
-    elevation_diff_to_previous = models.FloatField(null=True, blank=False)
-
-    def __str__(self) -> str:
-        return f"{self.number}"
-
-    class Meta:
-        ordering = ['segment__pk', 'number',]
-
-    def get_previous(self):
-        if self.number == 0:
-            # we are already on the start
-            return None
-
-        return GPXTrackSegmentPoint.objects.get(segment=self.segment, number=(self.number - 1))
-
-    def get_next(self):
-        try:
-            return GPXTrackSegmentPoint.objects.get(segment=self.segment, number=(self.number + 1))
-        except ObjectDoesNotExist:
-            return None
-
-
-class GPXFileUserSegmentSplit(TimeStampedModel):
-    name = models.CharField(null=True, blank=False)
-    gpx_file = models.ForeignKey("GPXFile", on_delete=models.CASCADE, related_name='user_segments')
-    point_start = models.ForeignKey("GPXTrackSegmentPoint", on_delete=models.CASCADE, related_name='user_splits_start')
-    point_end = models.ForeignKey("GPXTrackSegmentPoint", on_delete=models.CASCADE, related_name='user_splits_end')
-
-    distance = models.FloatField(null=True, blank=False)
-    total_ascent = models.FloatField(null=True, blank=False)
-    total_descent = models.FloatField(null=True, blank=False)
-
-    def __str__(self) -> str:
-        return f"{self.name}"
-
-    class Meta:
-        ordering = ['point_start__pk', 'id',]
-
-    def get_human_distance(self) -> str:
-        if self.distance:
-            d = int(self.distance / 1000)
-        else:
-            d = 0
-        return f"{d} km"
-
-    def get_total_ascent(self) -> int | None:
-        if self.total_ascent:
-            return int(self.total_ascent)
-
-        return None
-
-    def get_total_descent(self) -> int | None:
-        if self.total_descent:
-            return int(self.total_descent)
-
-        return None
-
-    def get_segment_id(self) -> int:
-        return self.point_start.segment.pk
-
-    def generate_gpx(self) -> str:
-        points = GPXTrackSegmentPoint.objects.filter(
-            segment__id=self.point_start.segment.pk,
-            number__gte=self.point_start.number,
-            number__lte=self.point_end.number,
-        )
-
-        gpx = GPX()
-        gpx.creator = 'GPX Tools by hggh'
-        gpx.name = self.name
-
-        track = Track()
-        track.name = self.name
-        segment = TrackSegment()
-
-        for point in points:
-            w = Waypoint()
-            w.lat = point.location.x
-            w.lon = point.location.y
-            w.ele = point.elevation
-
-            segment.points.append(w)
-
-        track.trksegs = [segment]
-        gpx.tracks = [track]
-
-        return gpx.to_string()
-
-    @staticmethod
-    def add_segment(gpx_file, segment_pk, point_number):
-        point = GPXTrackSegmentPoint.objects.get(
-            segment__id=segment_pk,
-            segment__track__gpx_file=gpx_file,
-            number=point_number,
-        )
-        update_splits = []
-
-        # check if there are any splits?
-        splits = GPXFileUserSegmentSplit.objects.filter(
-            gpx_file=gpx_file,
-            point_start__segment_id=segment_pk,
-
-        )
-        if len(splits) == 0:
-            s = GPXFileUserSegmentSplit(
-                gpx_file=gpx_file,
-                point_start=point.segment.points.all().first(),
-                point_end=point,
-            )
-            s.save()
-            update_splits.append(s)
-            s = GPXFileUserSegmentSplit(
-                gpx_file=gpx_file,
-                point_start=point,
-                point_end=point.segment.points.all().last(),
-            )
-            s.save()
-            update_splits.append(s)
-        else:
-            # we go back on the track and try to find a split start
-            start = GPXFileUserSegmentSplit.objects.filter(
-                    gpx_file=gpx_file,
-                    point_start__segment_id=segment_pk,
-                    point_start__number__lt=point.number,
-            ).order_by('point_start__number').last()
-
-            if start:
-                end_point = start.point_end
-
-                start.point_end = point
-                start.save()
-                update_splits.append(start)
-
-                s = GPXFileUserSegmentSplit(
-                    gpx_file=gpx_file,
-                    point_start=point,
-                    point_end=end_point,
-                )
-                s.save()
-                update_splits.append(s)
-
-        GPXFileUserSegmentSplit.update_segments(update_splits, gpx_file=gpx_file, segment_pk=segment_pk)
-
-    @staticmethod
-    def update_segments(update_splits, gpx_file, segment_pk):
-        for split in update_splits:
-            points = GPXTrackSegmentPoint.objects.filter(
-                segment__id=split.point_start.segment.id,
-                number__gte=split.point_start.number,
-                number__lte=split.point_end.number,
-            )
-            segment_total_ascent = 0
-            segment_total_descent = 0
-            total_distance = 0
-            priv_point = None
-
-            for p in points:
-                elevation_diff_to_previous = None
-
-                if priv_point is not None:
-                    if priv_point.elevation is not None and p.elevation is not None:
-                        elevation_diff_to_previous = priv_point.elevation - p.elevation
-
-                if priv_point is not None:
-                    total_distance += geopy_distance.distance(
-                        (priv_point.location.x, priv_point.location.y),
-                        (p.location.x, p.location.y)
-                    ).m
-                priv_point = p
-
-                if elevation_diff_to_previous is not None:
-                    if elevation_diff_to_previous > 0:
-                        segment_total_descent += elevation_diff_to_previous
-                    else:
-                        segment_total_ascent += elevation_diff_to_previous
-            split.total_ascent=abs(segment_total_ascent)
-            split.total_descent=abs(segment_total_descent)
-            split.distance=total_distance
-            split.save()
-
-        # name all splits
-        splits = GPXFileUserSegmentSplit.objects.filter(
-            gpx_file=gpx_file,
-            point_start__segment_id=segment_pk,
-
-        ).order_by('point_start__number')
-        split_number = 1
-        for split in splits:
-            split.name = f"Track {split_number}"
-            split.save()
-
-            split_number += 1
-
-        return None
-
-
-class GPXWayPointType(models.Model):
-    name = models.TextField(max_length=100, null=False, blank=False)
-    gpx_sym_name = models.CharField(max_length=100)
-    osm_name = models.TextField(max_length=100, null=False, blank=False)
-    osm_value = models.TextField(max_length=100, null=False, blank=False)
-    osm_query_type = models.CharField(max_length=10, null=False, blank=False, default='=')
-    around = models.IntegerField(null=False, blank=False)
-    around_max = models.IntegerField(null=False, blank=False)
-    around_duplicate = models.IntegerField(null=False, blank=False, default=3000)
-    marker_filename = models.CharField(max_length=100, null=True)
-    checked = models.BooleanField(null=False, default=True, blank=False)
-
-    def __str__(self) -> str:
-        return self.name
-
-    def get_json_data(self) -> dict:
-        data = {
-            "name": self.name,
-            "html_id": self.html_id(),
-            "marker_image_path": self.marker_image_path(),
-        }
-        return data
-
-    def html_id(self) -> str:
-        return "{}_{}".format(self.osm_name, self.osm_value).replace('|', '')
-
-    def marker_image_path(self) -> str:
-        return '/static/{}'.format(self.marker_filename)
-
-
-class GPXTrackWayPoint(TimeStampedModel):
-    gpx_file = models.ForeignKey("GPXFile", on_delete=models.CASCADE, related_name='waypoints')
-    waypoint_type = models.ForeignKey("GPXWayPointType", on_delete=models.CASCADE, related_name='waypoints')
-    name = models.CharField(max_length=100, null=False, blank=True)
-    tags = models.JSONField(default=dict)
-    hidden = models.BooleanField(null=False, blank=False, default=False)
-
-    location = models.PointField(default=Point(0.0, 0.0), db_index=True)
-    track_segment_point_nearby = models.ForeignKey("GPXTrackSegmentPoint", related_name='waypoints', on_delete=models.CASCADE)
-
-    class Meta:
-        ordering = ['track_segment_point_nearby__pk', 'id']
-
-    def get_json_data(self) -> dict:
-        # FIXME: als geoJSON!?
-        data = {
-            "id": self.pk,
-            "has_gpx_track_to": self.has_gpx_track_to(),
-            "lat": self.location.x,
-            "lon": self.location.y,
-            "class_name": self.get_marker_css_name(),
-            "url": self.get_url(),
-            "hidden": self.hidden,
-            "name": self.name,
-            "waypoint_type": {
-                "name": self.waypoint_type.name,
-                "html_id": self.waypoint_type.html_id(),
-                "marker_image_path": self.waypoint_type.marker_image_path(),
-
-            }
-        }
-        return data
-
-    def has_gpx_track_to(self) -> bool:
-        try:
-            if self.track_to_waypoint:
-                return True
-        except ObjectDoesNotExist:
-            pass
-        return False
-
-    def get_marker_css_name(self) -> str:
-        if self.is_camping_site() is True and self.get_url():
-            return 'marker_camping_with_url'
-        return 'marker_default'
-
-    def get_tags(self) -> dict:
-        tags = self.tags
-        tags.pop('contact:website', None)
-        tags.pop('phone', None)
-        tags.pop('website', None)
-        tags.pop('url', None)
-        tags.pop('name', None)
-
-        return tags
-
-    def get_phone(self) -> str | None:
-        if self.tags.get('phone', None):
-            try:
-                phonenumbers.parse(self.tags.get('phone'), None)
-                return self.tags.get('phone')
-            except Exception as e:
-                pass
-        return None
-
-    def get_url(self) -> str:
-        if self.tags.get('contact:website', None):
-            return self.tags.get('contact:website')
-
-        if self.tags.get('url', None):
-            return self.tags.get('url', None)
-
-        return self.tags.get('website', None)
-
-    def is_camping_site(self) -> bool:
-        if self.waypoint_type.osm_value == 'camp_site':
-            return True
-        return False
-
-    def is_hotel(self) -> bool:
-        if 'hotel' in self.waypoint_type.osm_value:
-            return True
-        return False
-
-
-class GPXTrackWayPointFromTrack(TimeStampedModel):
-    waypoint = models.OneToOneField("GPXTrackWayPoint", on_delete=models.CASCADE, related_name='track_to_waypoint')
-    geojson = models.JSONField(default=dict, null=False, blank=False)
-    away_kilometer = models.FloatField(null=False, blank=False, default=0.0)
-
-    def __str__(self) -> str:
-        return f"{self.away_kilometer}"
-
-    def get_download_url(self) -> str:
-        return f"/gpxtrack/{self.waypoint.gpx_file.slug}/download_gpx_track_to_waypoint/{self.waypoint.pk}"
-
-    def get_away_kilometer(self) -> int:
-        if self.away_kilometer > 1:
-            return int(self.away_kilometer)
-        else:
-            return 1
-
-    def get_geojson(self) -> dict:
-        fc = geojson.FeatureCollection(features=[])
-        fc.features.append(geojson.Feature(
-            geometry=geojson.LineString(coordinates=self.geojson),
-            properties={
-                'color': '#448137',
-                'weight': 3,
-                'opacity': 0.7,
-            }
-        ))
-
-        return fc
-
-    def get_gpx_track(self) -> str:
-        gpx = GPX()
-        gpx.creator = 'GPXViewr by hggh'
-        if self.waypoint.name != "":
-            name = f"to {self.waypoint.name}"
-        else:
-            name = f"to {self.waypoint.waypoint_type.osm_value}"
-        gpx.name = name
-
-        track = Track()
-        track.name = name
-
-        segment = TrackSegment()
-
-        for latlon in self.geojson:
-            w = Waypoint()
-            w.lat = latlon[1]
-            w.lon = latlon[0]
-
-            segment.points.append(w)
-
-        track.trksegs = [segment]
-        gpx.tracks = [track]
-
-        return gpx.to_string()
 
 
 @receiver(post_delete, sender=GPXFile)