+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 ### X. March 2025 [DRAFT]
   * Django DB Migration on Startup
+  * make contact:phone tag also as tel: link
+  * internal reorg of models into files
 
 ### 1. March 2025
   * handle Overpass Server Load too high - retry it after 4 seconds